Fix argument passing
The way arguments are passed with the plugin is not compatible with all
rebar providers/commands. Especially the release (relx) command does not
like getting it's arguments as a single list.

Commands like eunit are fine with that, but also do accept split
arguments. This fix changes argument passing to a splitted list of
strings in general. This way the plugin plays well with all commands.

This fix has been tested with the builtin commands, e.g. release, eunit
as well as with plugins, e.g. rebar3_proper.